[Rename] ReviewCard molecules로 이동
Resolves 장소 카드, 리뷰 카드 상세 페이지 이동 기능 추가 #26
